requirements.txt
requests>=2.25.1
selenium>=4.2.0
python-dotenv>=0.19.0
schedule>=1.1.0
wechatpy>=1.8.18
pycryptodome>=3.16.0
pillow>=9.2.0
openai>=1.3.0
beautifulsoup4>=4.9.3
lxml>=4.9.1
webdriver-manager>=3.8.4
playwright>=1.30.0
tqdm>=4.61.1 